Upgrade terraform-provider-azurerm to v3.74.0 (#1359)
This PR was generated via `$ upgrade-provider pulumi/pulumi-azure
--kind=all`.

---

- Upgrading terraform-provider-azurerm from 3.72.0  to 3.74.0.
	Fixes #1355
	Fixes #1354
- Upgrading pulumi-terraform-bridge from v3.59.0 to v3.60.1.
- Upgrading pulumi/terraform-plugin-sdk from v2.27.0 to 2.29.0.

+++ sdk/dotnet/AppConfiguration/ConfigurationFeature.cs
@@ -118,10 +118,10 @@ namespace Pulumi.Azure.AppConfiguration
         public Output<string> Name { get; private set; } = null!;
 
         /// <summary>
-        /// A list of one or more numbers representing the value of the percentage required to enable this feature.
+        /// A number representing the value of the percentage required to enable this feature.
         /// </summary>
         [Output("percentageFilterValue")]
-        public Output<int?> PercentageFilterValue { get; private set; } = null!;
+        public Output<double?> PercentageFilterValue { get; private set; } = null!;
 
         /// <summary>
         /// A mapping of tags to assign to the resource.
@@ -233,10 +233,10 @@ namespace Pulumi.Azure.AppConfiguration
         public Input<string>? Name { get; set; }
 
         /// <summary>
-        /// A list of one or more numbers representing the value of the percentage required to enable this feature.
+        /// A number representing the value of the percentage required to enable this feature.
         /// </summary>
         [Input("percentageFilterValue")]
-        public Input<int>? PercentageFilterValue { get; set; }
+        public Input<double>? PercentageFilterValue { get; set; }
 
         [Input("tags")]
         private InputMap<string>? _tags;
@@ -328,10 +328,10 @@ namespace Pulumi.Azure.AppConfiguration
         public Input<string>? Name { get; set; }
 
         /// <summary>
-        /// A list of one or more numbers representing the value of the percentage required to enable this feature.
+        /// A number representing the value of the percentage required to enable this feature.
         /// </summary>
         [Input("percentageFilterValue")]
-        public Input<int>? PercentageFilterValue { get; set; }
+        public Input<double>? PercentageFilterValue { get; set; }
 
         [Input("tags")]
         private InputMap<string>? _tags;
